update controllers to return more consistant data, sends error object on error and data object on success

+++ src/controllers/userControllers.ts
@@ -1,40 +1,56 @@
 import { NextFunction, Request, Response } from 'express'
 import { User } from '../models/models-exports'
+import { Types } from 'mongoose'
 
 // GET Requests
-export const get_user = async (req: Request, res: Response): Promise<void> => {
-	const { id } = req.params
-
-	try {
-		const user = await User.findOne({ _id: id })
-
-		if (!user) {
-			res.status(404).json({ error: 'No user found with this id.' })
-		} else {
-			res.status(200).json({ user: { email: user?.email } })
-		}
-	} catch (error) {
-		res.status(404).json({ error })
-	}
-}
-
-export const get_current_user = async (
+export const get_user = async (
 	req: Request,
 	res: Response
-): Promise<void> => {
-	let user = req.user as any
+): Promise<void | Response> => {
+	const { id } = req.params
 
-	if (req.user !== undefined) {
-		res.json({
+	if (!Types.ObjectId.isValid(id)) {
+		return res.status(404).json({ error: 'Invalid id' })
+	}
+
+	const user = await User.findOne({ _id: id })
+
+	if (!user) {
+		return res.status(404).json({ error: 'No user found with that id' })
+	}
+
+	res.status(200).json({
+		data: {
 			user: {
 				email: user?.email,
 				id: user?._id,
 				role: user?.role,
 				username: user?.username,
 			},
+		},
+	})
+}
+
+export const get_current_user = async (
+	req: Request,
+	res: Response
+): Promise<void> => {
+	// req.user is type any to access the properties
+	let user = req.user as any
+
+	if (req.user) {
+		res.status(200).json({
+			data: {
+				user: {
+					email: user?.email,
+					id: user?._id,
+					role: user?.role,
+					username: user?.username,
+				},
+			},
 		})
 	} else {
-		res.json({ user })
+		res.status(200).json({ data: req.user })
 	}
 }
 
@@ -42,12 +58,26 @@ export const get_current_user = async (
 export const signup_user = async (
 	req: Request,
 	res: Response
-): Promise<void> => {
-	try {
-		const user = await User.create(req.body)
-		res.status(200).json({ message: 'User successfully created' })
-	} catch (error) {
-		res.status(400).json({ error })
+): Promise<void | Response> => {
+	const checkIfUser = await User.findOne({ email: req.body.email })
+
+	if (checkIfUser) {
+		return res.status(400).json({ error: 'This email is already registered ' })
+	} else {
+		User.create(req.body)
+			.then((user) =>
+				res.status(200).json({
+					data: {
+						user: {
+							email: user.email,
+							id: user._id,
+							role: user.role,
+							username: user.username,
+						},
+					},
+				})
+			)
+			.catch(() => res.status(400).json({ error: 'Unable to sign up user' }))
 	}
 }
 
@@ -59,13 +89,17 @@ export const login_user = async (
 
 	if (req.user) {
 		res.status(200).json({
-			email: user.email,
-			id: user._id,
-			role: user.role,
-			username: user.username,
+			data: {
+				user: {
+					email: user?.email,
+					id: user?._id,
+					role: user?.role,
+					username: user?.username,
+				},
+			},
 		})
 	} else {
-		res.status(400).json({ user })
+		res.status(200).json({ data: req.user })
 	}
 }
 
@@ -76,7 +110,7 @@ export const logout_user = async (
 ): Promise<void | NextFunction> => {
 	req.logout((err) => {
 		if (err) return next(err)
-		res.status(200).json(req.user)
+		res.status(200).json({ data: req.user })
 	})
 }
 
@@ -84,18 +118,18 @@ export const logout_user = async (
 export const delete_user = async (
 	req: Request,
 	res: Response
-): Promise<void> => {
+): Promise<void | Response> => {
 	const { id } = req.params
 
-	try {
-		const user = await User.findOneAndDelete({ _id: id })
-
-		if (!user) {
-			res.status(404).json({ error: 'No user found with this id.' })
-		} else {
-			res.status(200).json(user)
-		}
-	} catch (error) {
-		res.status(404).json({ error })
+	if (!Types.ObjectId.isValid(id)) {
+		return res.status(404).json({ error: 'Invalid id' })
 	}
+
+	const user = await User.findOneAndDelete({ _id: id })
+
+	if (!user) {
+		return res.status(404).json({ error: 'No user found with this id' })
+	}
+
+	res.status(200).json({ data: user })
 }
